+++ react-shopping-cart/src/index.jsx
@@ -3,26 +3,31 @@ import ReactDOM from 'react-dom/client';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
+import { persistStore } from 'redux-persist';
+import { PersistGate } from 'redux-persist/integration/react';
 import { ThemeProvider } from 'styled-components';
 import ProductList from 'pages/ProductListPage/ProductList.page';
-import shoppingCartReducer from 'reducers';
+import rootReducer from 'reducers';
 import GlobalStyle from './styles/GlobalStyle';
 import theme from './styles/theme';
 
-const store = createStore(shoppingCartReducer);
+const store = createStore(rootReducer);
+const persistor = persistStore(store);
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
     <Provider store={store}>
-      <GlobalStyle />
-      <BrowserRouter>
-        <ThemeProvider theme={theme}>
-          <Routes>
-            <Route path="/" element={<ProductList />} />
-          </Routes>
-        </ThemeProvider>
-      </BrowserRouter>
+      <PersistGate persistor={persistor}>
+        <GlobalStyle />
+        <BrowserRouter>
+          <ThemeProvider theme={theme}>
+            <Routes>
+              <Route path="/" element={<ProductList />} />
+            </Routes>
+          </ThemeProvider>
+        </BrowserRouter>
+      </PersistGate>
     </Provider>
   </React.StrictMode>
 );
